Minor refactoring
Export unexported base class for exported classes
Replace int with std::size_t for size related functions
Silence possible data loss warning with explicit cast


Former-commit-id: 482e26331000438a06749379664888bb50b2d196

+++ src/IME/core/loop/Engine.cpp
@@ -157,7 +157,7 @@ namespace ime {
         while (window_.isOpen() && isRunning_ && !statesManager_.isEmpty()) {
             if (onFrameStart_)
                 onFrameStart_();
-            now = clock.getElapsedTimeInSeconds();
+            now = static_cast<float>(clock.getElapsedTimeInSeconds());
             deltaTime = now - prevTime;
             prevTime = now;
             accumulator += deltaTime;
